+++ src/App.js
@@ -1,12 +1,13 @@
-// src/App.js
 import React from "react";
 import { BrowserRouter as Router, Routes, Route, Navigate } from "react-router-dom";
 import { AuthProvider, useAuth } from "./utils/Auth/AuthContext";
 import LoginForm from "./pages/LoginFormPage/loginFormPage";
 import BuyerDbPage from "./pages/BuyerDbPage/buyerDbPage";
 import SellerDbPage from "./pages/SellerDbPage/sellerDbPage";
-import ProtectedRoute from "./utils/ProtectedRoute/ProtectedRoute";
 import AdminPage from "./pages/Admin/adminPage";
+import Chat from "./components/chat/chat"; // Import Chat component
+import ChatPage from "./components/chat/ChatPage"; // Import ChatPage component
+import ProtectedRoute from "./utils/ProtectedRoute/ProtectedRoute";
 
 const AppRoutes = () => {
   const { isLoggedIn, userType } = useAuth();
@@ -34,7 +35,7 @@ const AppRoutes = () => {
           </ProtectedRoute>
         }
       />
-       <Route
+      <Route
         path="/admin/*"
         element={
           <ProtectedRoute allowedRoles={["admin"]}>
@@ -42,8 +43,9 @@ const AppRoutes = () => {
           </ProtectedRoute>
         }
       />
+      <Route path="/chat" element={<Chat />} /> {/* Route for displaying chat rooms */}
+      <Route path="/chat/:roomId" element={<ChatPage />} /> {/* Route for individual chat room */}
     </Routes>
-    
   );
 };
