+++ src/components/BuyerDb/buyerDb.jsx
@@ -1,4 +1,5 @@
-import React, { useState } from 'react';
+// BuyerDb.js
+import React, { useState, useEffect } from 'react';
 import { useAuth } from '../../utils/Auth/AuthContext';
 import Button from '@mui/material/Button';
 import { useNavigate } from 'react-router-dom';
@@ -6,9 +7,8 @@ import Sidebar from '../SideBar/sidebar'; // Ensure correct import path to Sideb
 import Paper from '@mui/material/Paper';
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
-import Chat from '../Chat'; // Ensure correct import path to Chat component
 import { keyframes } from '@emotion/react';
-
+import ManageProposals from "./features/ManageProposals";
 // Import the necessary features/components
 import C2I from './features/C2I';
 import I2C from './features/I2C';
@@ -16,13 +16,13 @@ import C2C from './features/C2C';
 import PAYAPP from './features/payApp';
 import Help from './features/help';
 import Contact from './features/contact';
+import ChatRooms from '../chat/chat';  // New component to list chat rooms
 
 const BuyerDb = () => {
   const { currentUser, logout } = useAuth();
   const navigate = useNavigate();
   const [selectedCategory, setSelectedCategory] = useState('');
-  const [selectedProposal, setSelectedProposal] = useState(null); // State to hold selected proposal
-console.log(setSelectedProposal)
+
   const handleLogout = () => {
     logout();
     navigate('/');
@@ -33,11 +33,12 @@ console.log(setSelectedProposal)
     { name: 'I2C' },
     { name: 'C2C' },
     { name: 'PAYAPP' },
+    { name: 'ManageProposals' },
+    { name: 'Chat Rooms' },  // New menu option
     { name: 'Help' },
     { name: 'Contact' },
   ];
 
- 
   const fadeIn = keyframes`
     from {
       opacity: 0;
@@ -46,14 +47,6 @@ console.log(setSelectedProposal)
       opacity: 1;
     }
   `;
-
-  const renderChat = () => {
-    if (selectedProposal) {
-      const roomId = `Chat Room - ${selectedProposal.id}`; // Assuming selectedProposal has an id
-      return <Chat currentUser={currentUser} roomId={roomId} />;
-    }
-    return null;
-  };
 
   const renderContent = () => {
     switch (selectedCategory) {
@@ -65,6 +58,10 @@ console.log(setSelectedProposal)
         return <C2C />;
       case 'PAYAPP':
         return <PAYAPP />;
+      case 'ManageProposals':
+        return <ManageProposals />;
+      case 'Chat Rooms':  // New case for chat rooms
+        return <ChatRooms userRole="buyer" />;
       case 'Help':
         return <Help />;
       case 'Contact':
@@ -78,10 +75,10 @@ console.log(setSelectedProposal)
                 margin: '10px', 
                 width: '45%', 
                 backgroundColor: '#0074D9', 
-                color: '#ffffff',
+                color: '#ffffff' 
               }}>
                 <Typography variant="h6" gutterBottom>All Deals</Typography>
-                <Typography variant="body1">Animated content for all deals...</Typography>
+                <Typography variant="body1">Content for managing deals...</Typography>
               </Paper>
 
               <Paper elevation={5} style={{ 
@@ -89,7 +86,7 @@ console.log(setSelectedProposal)
                 margin: '10px', 
                 width: '45%', 
                 backgroundColor: '#2ECC40', 
-                color: '#ffffff',
+                color: '#ffffff' 
               }}>
                 <Typography variant="h6" gutterBottom>Successful Deals</Typography>
                 <Typography variant="body1">Content for successful deals...</Typography>
@@ -102,7 +99,7 @@ console.log(setSelectedProposal)
                 margin: '10px', 
                 width: '45%', 
                 backgroundColor: '#FF4136', 
-                color: '#ffffff',
+                color: '#ffffff' 
               }}>
                 <Typography variant="h6" gutterBottom>Unsuccessful Deals</Typography>
                 <Typography variant="body1">Content for unsuccessful deals...</Typography>
@@ -113,27 +110,24 @@ console.log(setSelectedProposal)
                 margin: '10px', 
                 width: '45%', 
                 backgroundColor: '#FF851B', 
-                color: '#ffffff',
+                color: '#ffffff' 
               }}>
                 <Typography variant="h6" gutterBottom>User Level</Typography>
                 <Typography variant="body1">Content for user level...</Typography>
               </Paper>
             </Box>
 
-            <Paper elevation={5} style={{ 
-              padding: '20px', 
-              margin: '10px', 
-              width: '45%', 
-              backgroundColor: '#FFDC00', 
-              color: '#000000',
-            }}>
-              <Typography variant="h6" gutterBottom>Transaction History</Typography>
-              <Typography variant="body1">Content for transaction history...</Typography>
-            </Paper>
-
-            <Box>
-              {/* Render the Chat component */}
-              {renderChat()}
+            <Box style={{ display: 'flex', justifyContent: 'center', marginTop: '20px' }}>
+              <Paper elevation={5} style={{ 
+                padding: '20px', 
+                margin: '10px', 
+                width: '45%', 
+                backgroundColor: '#FFDC00', 
+                color: '#000000' 
+              }}>
+                <Typography variant="h6" gutterBottom>Transaction History</Typography>
+                <Typography variant="body1">Content for transaction history...</Typography>
+              </Paper>
             </Box>
           </div>
         );
@@ -142,20 +136,20 @@ console.log(setSelectedProposal)
 
   return (
     <div>
-      <Sidebar leftOptions={buyerOptions}  onItemClick={setSelectedCategory} />
+      <Sidebar leftOptions={buyerOptions} onItemClick={setSelectedCategory} />
 
-      <nav style={{ 
-        width: '100%', 
-        backgroundColor: '#001f3f', 
-        color: '#ffffff', 
-        padding: '10px', 
-        display: 'flex', 
-        justifyContent: 'space-around', 
+      <nav style={{
+        width: '100%',
+        backgroundColor: '#001f3f',
+        color: '#ffffff',
+        padding: '10px',
+        display: 'flex',
+        justifyContent: 'space-around',
         alignItems: 'center',
-        position: 'fixed', 
-        top: '0', 
-        zIndex: '1000', 
-        borderBottom: '2px solid #ffffff', 
+        position: 'fixed',
+        top: '0',
+        zIndex: '1000',
+        borderBottom: '2px solid #ffffff',
         boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)',
       }}>
         <h1 style={{ margin: '0', fontSize: '1.5rem' }}>Buyer Dashboard</h1>
@@ -164,15 +158,14 @@ console.log(setSelectedProposal)
         </Button>
       </nav>
 
-      <Box style={{ 
-        display: 'flex', 
-        justifyContent: 'center', 
-        alignItems: 'center', 
-        flexDirection: 'column', 
-        marginTop: '100px', 
-        animation: `${fadeIn} 1s ease-in`, 
+      <Box style={{
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+        flexDirection: 'column',
+        marginTop: '100px',
+        animation: `${fadeIn} 1s ease-in`,
       }}>
-        {/* Render the content based on selectedCategory */}
         {renderContent()}
       </Box>
     </div>
